refactor(commands): extract assorted commands into extras.rs
