+++ src-rust/crates/commands/src/extras.rs
@@ -1,0 +1,268 @@
+// Assorted commands: `/advisor`, `/install-slack-app`, `/fast`, `/feedback`, `/color` (full).
+//
+// Extracted from lib.rs (issue #232). Behavior-preserving move.
+
+use super::*;
+use async_trait::async_trait;
+
+pub struct AdvisorCommand;
+pub struct InstallSlackAppCommand;
+pub struct FastCommand;
+pub struct FeedbackCommand;
+pub struct ColorSetCommand;
+
+// ---- /advisor ------------------------------------------------------------
+
+#[async_trait]
+impl SlashCommand for AdvisorCommand {
+    fn name(&self) -> &str { "advisor" }
+    fn description(&self) -> &str { "Set or unset the server-side advisor model" }
+    fn help(&self) -> &str {
+        "Usage: /advisor [<model>|off|unset]\n\n\
+         Sets the advisor model used for server-side suggestions.\n\
+         Examples:\n\
+           /advisor claude-opus-4-6   — set advisor model\n\
+           /advisor off               — disable the advisor\n\
+           /advisor                   — show current advisor setting"
+    }
+
+    async fn execute(&self, args: &str, _ctx: &mut CommandContext) -> CommandResult {
+        let arg = args.trim();
+        let settings_dir = claurst_core::config::Settings::config_dir();
+        let settings_path = settings_dir.join("settings.json");
+
+        // Read or create settings JSON
+        let mut settings_val: serde_json::Value = settings_path
+            .exists()
+            .then(|| std::fs::read_to_string(&settings_path).ok())
+            .flatten()
+            .and_then(|s| serde_json::from_str(&s).ok())
+            .unwrap_or_else(|| serde_json::json!({}));
+
+        match arg {
+            "" => {
+                let current = settings_val
+                    .get("advisorModel")
+                    .and_then(|v| v.as_str())
+                    .unwrap_or("(not set)");
+                CommandResult::Message(format!("Advisor model: {current}"))
+            }
+            "off" | "unset" | "none" => {
+                settings_val
+                    .as_object_mut()
+                    .map(|m| m.remove("advisorModel"));
+                if let Ok(json) = serde_json::to_string_pretty(&settings_val) {
+                    let _ = std::fs::write(&settings_path, json);
+                }
+                CommandResult::Message("Advisor model unset.".to_string())
+            }
+            model => {
+                // Basic validation: must look like a model identifier
+                if model.starts_with("claude-") || model.contains('/') {
+                    settings_val["advisorModel"] = serde_json::Value::String(model.to_string());
+                    if let Ok(json) = serde_json::to_string_pretty(&settings_val) {
+                        let _ = std::fs::write(&settings_path, json);
+                    }
+                    CommandResult::Message(format!("Advisor model set to: {model}"))
+                } else {
+                    CommandResult::Message(format!(
+                        "Unknown model '{model}'. Model IDs should start with 'claude-'.\n\
+                         Use /model to see available models."
+                    ))
+                }
+            }
+        }
+    }
+}
+
+// ---- /install-slack-app --------------------------------------------------
+
+#[async_trait]
+impl SlashCommand for InstallSlackAppCommand {
+    fn name(&self) -> &str { "install-slack-app" }
+    fn description(&self) -> &str { "Install the Claurst Slack integration" }
+    fn help(&self) -> &str {
+        "Usage: /install-slack-app\n\n\
+         Opens instructions for installing the Claurst Slack app.\n\
+         Requires a Claurst for Enterprise subscription."
+    }
+
+    async fn execute(&self, _args: &str, _ctx: &mut CommandContext) -> CommandResult {
+        CommandResult::Message(
+            "Claurst Slack Integration\n\
+             ─────────────────────────────\n\
+             To install Claurst in Slack:\n\n\
+             1. Ensure you have a Claurst for Enterprise subscription\n\
+             2. Visit your Anthropic Console → Integrations → Slack\n\
+             3. Click \"Add to Slack\" and authorize the app\n\
+             4. Invite @Claurst to any channel with: /invite @Claurst\n\n\
+             In Slack, you can then:\n\
+             • Mention @Claurst to ask questions in any channel\n\
+             • Use /claude for direct commands\n\
+             • Share code snippets for review\n\n\
+             See: https://docs.anthropic.com/claude-code/slack"
+                .to_string(),
+        )
+    }
+}
+
+// ---- /fast (/speed) ------------------------------------------------------
+
+#[async_trait]
+impl SlashCommand for FastCommand {
+    fn name(&self) -> &str { "fast" }
+    fn aliases(&self) -> Vec<&str> { vec!["speed"] }
+    fn description(&self) -> &str { "Toggle fast mode (uses a faster/cheaper model)" }
+    fn help(&self) -> &str {
+        "Usage: /fast [on|off]\n\n\
+         Fast mode switches to the active provider's smaller, faster model\n\
+         for quick responses. Toggle without argument to switch.\n\
+         The setting is persisted to ~/.claurst/ui-settings.json."
+    }
+
+    async fn execute(&self, args: &str, ctx: &mut CommandContext) -> CommandResult {
+        let current = load_ui_settings();
+        let currently_on = current.fast_mode.unwrap_or(false);
+
+        let enable = match args.trim() {
+            "on" | "enable" | "true" | "1" => true,
+            "off" | "disable" | "false" | "0" => false,
+            "" => !currently_on,
+            other => {
+                return CommandResult::Error(format!(
+                    "Unknown argument '{}'. Use: /fast [on|off]",
+                    other
+                ))
+            }
+        };
+
+        if let Err(e) = mutate_ui_settings(|s| s.fast_mode = Some(enable)) {
+            return CommandResult::Error(format!("Failed to save setting: {}", e));
+        }
+
+        let provider_id = ctx.config.selected_provider_id();
+        let fast_model = resolve_fast_model_id(&ctx.config);
+        let normal_model = stripped_model_for_provider(
+            provider_id,
+            ctx.config.effective_model(),
+        )
+        .to_string();
+
+        if enable {
+            let mut new_config = ctx.config.clone();
+            new_config.model = Some(canonical_model_for_provider(provider_id, &fast_model));
+            CommandResult::ConfigChangeMessage(
+                new_config,
+                format!(
+                    "Fast mode ON. Using {} for quicker, cheaper responses.\n\
+                     Use /fast off to return to {}.",
+                    fast_model, normal_model
+                ),
+            )
+        } else {
+            let mut new_config = ctx.config.clone();
+            // Restore default / saved model
+            new_config.model = None;
+            let restored_model = stripped_model_for_provider(
+                provider_id,
+                new_config.effective_model(),
+            )
+            .to_string();
+            CommandResult::ConfigChangeMessage(
+                new_config,
+                format!(
+                    "Fast mode OFF. Restored to default model ({}).",
+                    restored_model
+                ),
+            )
+        }
+    }
+}
+
+// ---- /feedback (standalone, supplements BugCommand alias) ----------------
+
+#[async_trait]
+impl SlashCommand for FeedbackCommand {
+    fn name(&self) -> &str { "report" }
+    fn aliases(&self) -> Vec<&str> { vec![] }
+    fn description(&self) -> &str { "Open the GitHub issues page to report a bug or request a feature" }
+    fn hidden(&self) -> bool { true } // surfaced via BugCommand alias; hidden to avoid duplicate
+    fn help(&self) -> &str {
+        "Usage: /report [description]\n\n\
+         Opens the GitHub issues tracker. If a description is provided,\n\
+         it is shown as a suggested pre-fill for the issue body."
+    }
+
+    async fn execute(&self, args: &str, _ctx: &mut CommandContext) -> CommandResult {
+        let url = "https://github.com/anthropics/claude-code/issues/new";
+        let report = args.trim();
+        let display_url = if report.is_empty() {
+            url.to_string()
+        } else {
+            // Append as a body query param
+            format!(
+                "{}?body={}",
+                url,
+                urlencoding::encode(report)
+            )
+        };
+
+        match open_with_system(&display_url) {
+            Ok(_) => CommandResult::Message(format!("Opened issue tracker: {}", url)),
+            Err(_) => CommandResult::Message(format!(
+                "Please visit {} to submit a report.",
+                url
+            )),
+        }
+    }
+}
+
+// ---- /color (full implementation) ----------------------------------------
+
+#[async_trait]
+impl SlashCommand for ColorSetCommand {
+    fn name(&self) -> &str { "color-set" }
+    fn hidden(&self) -> bool { true }
+    fn description(&self) -> &str { "Internal: set prompt color — use /color instead" }
+
+    async fn execute(&self, args: &str, _ctx: &mut CommandContext) -> CommandResult {
+        let color = args.trim();
+        if color.is_empty() {
+            let current = load_ui_settings();
+            return CommandResult::Message(format!(
+                "Current prompt color: {}\n\
+                 Use /color <name|#RRGGBB|default> to change it.\n\n\
+                 Named colors: red, green, blue, yellow, cyan, magenta, white, orange, purple",
+                current.prompt_color.as_deref().unwrap_or("default"),
+            ));
+        }
+
+        let normalized = if color == "default" {
+            None
+        } else {
+            // Validate hex or named color
+            let known_colors = [
+                "red", "green", "blue", "yellow", "cyan", "magenta",
+                "white", "orange", "purple", "pink", "gray", "grey",
+            ];
+            let is_hex = color.starts_with('#') && (color.len() == 4 || color.len() == 7)
+                && color[1..].chars().all(|c| c.is_ascii_hexdigit());
+            if !is_hex && !known_colors.contains(&color.to_lowercase().as_str()) {
+                return CommandResult::Error(format!(
+                    "Unknown color '{}'. Use a color name (red, green, …) or a hex code (#RGB or #RRGGBB).",
+                    color
+                ));
+            }
+            Some(color.to_string())
+        };
+
+        match mutate_ui_settings(|s| s.prompt_color = normalized.clone()) {
+            Ok(_) => CommandResult::Message(format!(
+                "Prompt color set to {}.\n\
+                 Restart the REPL for the change to take effect.",
+                normalized.as_deref().unwrap_or("default")
+            )),
+            Err(e) => CommandResult::Error(format!("Failed to save color: {}", e)),
+        }
+    }
+}
